+++ src/App.tsx
@@ -18,35 +18,28 @@ import Settings from './components/settings';
 import Template from './components/template';
 
 
+
 import {
-  getStatusIndicatorData,
+  renderNameColumn,
+  renderLastRunColumn,
+  renderDateColumn,
+  renderType,
   IPipelineItem,
-  pipelineItems,
-  ReleaseType,
-  ReleaseTypeText,
+  pipelineItems
 } from "./data/TableData";
 
 import {
   ColumnMore,
   ITableColumn,
-  SimpleTableCell,
   Table,
-  TwoLineTableCell,
   ColumnSorting,
   SortOrder,
   sortItems,
 } from "azure-devops-ui/Table";
-import { Ago } from "azure-devops-ui/Ago";
-import { Duration } from "azure-devops-ui/Duration";
-import { Tooltip } from "azure-devops-ui/TooltipEx";
-import { css } from "azure-devops-ui/Util";
+
 import { ArrayItemProvider } from "azure-devops-ui/Utilities/Provider";
 import { ObservableValue } from "azure-devops-ui/Core/Observable";
 import { Observer } from "azure-devops-ui/Observer";
-import { Icon, IIconProps } from "azure-devops-ui/Icon";
-import { Link } from "azure-devops-ui/Link";
-import { Status, StatusSize } from "azure-devops-ui/Status";
-
 
 
 interface IAppState {
@@ -64,19 +57,16 @@ class App extends React.Component<{}, IAppState>  {
     };
   }
 
-  render()  {
+  render() {
     return (
       <Page>
         <CustomHeader className="bolt-header-with-commandbar">
           <HeaderTitleArea>
             <HeaderTitleRow>
               <HeaderTitle className="text-ellipsis" titleSize={TitleSize.Large}>
-                #3: Add new header to sample site
+                Stack board
               </HeaderTitle>
             </HeaderTitleRow>
-            <HeaderDescription>
-              Validation of 405810 triggered today at 6:27 pm
-            </HeaderDescription>
           </HeaderTitleArea>
           <ButtonGroup>
             <Button text="Create" iconProps={{ iconName: "Add" }} primary={true}
@@ -93,7 +83,7 @@ class App extends React.Component<{}, IAppState>  {
           <Card
             className="flex-grow bolt-table-card"
             contentProps={{ contentPadding: false }}
-            titleProps={{ text: "All pipelines" }}
+            titleProps={{ text: "All projects" }}
           >
             <Observer itemProvider={this.itemProvider}>
               {(observableProps: { itemProvider: ArrayItemProvider<IPipelineItem> }) => (
@@ -114,7 +104,6 @@ class App extends React.Component<{}, IAppState>  {
         </div>
 
         <Settings show={this.state.settingsExpanded} onDismiss={() => this.setState({ settingsExpanded: false })} />
-
         <Template show={this.state.createExpanded} onDismiss={() => this.setState({ createExpanded: false })} />
 
       </Page>
@@ -124,7 +113,7 @@ class App extends React.Component<{}, IAppState>  {
   columns: ITableColumn<IPipelineItem>[] = [
     {
       id: "name",
-      name: "Pipeline",
+      name: "Project",
       renderCell: renderNameColumn,
       readonly: true,
       sortProps: {
@@ -136,8 +125,15 @@ class App extends React.Component<{}, IAppState>  {
     {
       className: "pipelines-two-line-cell",
       id: "lastRun",
-      name: "Last run",
+      name: "Settings",
       renderCell: renderLastRunColumn,
+      width: -33,
+    },
+    {
+      id: "type",
+      ariaLabel: "Time and duration",
+      readonly: true,
+      renderCell: renderType,
       width: -33,
     },
     {
@@ -147,12 +143,13 @@ class App extends React.Component<{}, IAppState>  {
       renderCell: renderDateColumn,
       width: -33,
     },
+    
     new ColumnMore(() => {
       return {
         id: "sub-menu",
         items: [
-          { id: "submenu-one", text: "SubMenuItem 1" },
-          { id: "submenu-two", text: "SubMenuItem 2" },
+          { id: "download", text: "Download" },
+          { id: "delete", text: "Delete" },
         ],
       };
     }),
@@ -168,169 +165,17 @@ class App extends React.Component<{}, IAppState>  {
         sortItems(
           columnIndex,
           proposedSortOrder,
-          this.sortFunctions,
+          [
+            (item1: IPipelineItem, item2: IPipelineItem) => {
+              return item1.name.localeCompare(item2.name);
+            }
+          ],
           this.columns,
           pipelineItems
         )
       );
     }
   );
-
-  sortFunctions = [
-    // Sort on Name column
-    (item1: IPipelineItem, item2: IPipelineItem) => {
-      return item1.name.localeCompare(item2.name);
-    },
-  ];
-
 }
-
-
-function renderNameColumn(
-  rowIndex: number,
-  columnIndex: number,
-  tableColumn: ITableColumn<IPipelineItem>,
-  tableItem: IPipelineItem
-): JSX.Element {
-  return (
-    <SimpleTableCell
-      columnIndex={columnIndex}
-      tableColumn={tableColumn}
-      key={"col-" + columnIndex}
-      contentClassName="fontWeightSemiBold font-weight-semibold fontSizeM font-size-m scroll-hidden"
-    >
-      <Status
-        {...getStatusIndicatorData(tableItem.status).statusProps}
-        className="icon-large-margin"
-        size={StatusSize.l}
-      />
-      <div className="flex-row scroll-hidden">
-        <Tooltip overflowOnly={true}>
-          <span className="text-ellipsis">{tableItem.name}</span>
-        </Tooltip>
-      </div>
-    </SimpleTableCell>
-  );
-}
-
-function renderLastRunColumn(
-  rowIndex: number,
-  columnIndex: number,
-  tableColumn: ITableColumn<IPipelineItem>,
-  tableItem: IPipelineItem
-): JSX.Element {
-  const { prName, prId, releaseType, branchName } = tableItem.lastRunData;
-  const text = "#" + prId + " \u00b7 " + prName;
-  const releaseTypeText = ReleaseTypeText({ releaseType: releaseType });
-  const tooltip = `${releaseTypeText} from ${branchName} branch`;
-  return (
-    <TwoLineTableCell
-      className="bolt-table-cell-content-with-inline-link no-v-padding"
-      key={"col-" + columnIndex}
-      columnIndex={columnIndex}
-      tableColumn={tableColumn}
-      line1={
-        <span className="flex-row scroll-hidden">
-          <Tooltip text={text} overflowOnly>
-            <Link
-              className="fontSizeM font-size-m text-ellipsis bolt-table-link bolt-table-inline-link"
-              excludeTabStop
-              href="#pr"
-            >
-              {text}
-            </Link>
-          </Tooltip>
-        </span>
-      }
-      line2={
-        <Tooltip text={tooltip} overflowOnly>
-          <span className="fontSize font-size secondary-text flex-row flex-center text-ellipsis">
-            {ReleaseTypeIcon({ releaseType: releaseType })}
-            <span className="text-ellipsis" key="release-type-text" style={{ flexShrink: 10000 }}>
-              {releaseTypeText}
-            </span>
-            <Link
-              className="monospaced-text text-ellipsis flex-row flex-center bolt-table-link bolt-table-inline-link"
-              excludeTabStop
-              href="#branch"
-            >
-              {Icon({
-                className: "icon-margin",
-                iconName: "OpenSource",
-                key: "branch-name",
-              })}
-              <span className="text-ellipsis">
-                {branchName}
-              </span>
-            </Link>
-          </span>
-        </Tooltip>
-      }
-    />
-  );
-}
-
-function renderDateColumn(
-  rowIndex: number,
-  columnIndex: number,
-  tableColumn: ITableColumn<IPipelineItem>,
-  tableItem: IPipelineItem
-): JSX.Element {
-  return (
-    <TwoLineTableCell
-      key={"col-" + columnIndex}
-      columnIndex={columnIndex}
-      tableColumn={tableColumn}
-      line1={WithIcon({
-        className: "fontSize font-size",
-        iconProps: { iconName: "Calendar" },
-        children: (
-          <Ago date={tableItem.lastRunData.startTime} /*format={AgoFormat.Extended}*/ />
-        ),
-      })}
-      line2={WithIcon({
-        className: "fontSize font-size bolt-table-two-line-cell-item",
-        iconProps: { iconName: "Clock" },
-        children: (
-          <Duration
-            startDate={tableItem.lastRunData.startTime}
-            endDate={tableItem.lastRunData.endTime}
-          />
-        ),
-      })}
-    />
-  );
-}
-
-function WithIcon(props: {
-  className?: string;
-  iconProps: IIconProps;
-  children?: React.ReactNode;
-}) {
-  return (
-    <div className={css(props.className, "flex-row flex-center")}>
-      {Icon({ ...props.iconProps, className: "icon-margin" })}
-      {props.children}
-    </div>
-  );
-}
-
-function ReleaseTypeIcon(props: { releaseType: ReleaseType }) {
-  let iconName: string = "";
-  switch (props.releaseType) {
-    case ReleaseType.PrAutomated:
-      iconName = "BranchPullRequest";
-      break;
-    default:
-      iconName = "Tag";
-  }
-
-  return Icon({
-    className: "bolt-table-inline-link-left-padding icon-margin",
-    iconName: iconName,
-    key: "release-type",
-  });
-}
-
 
 export default App;
